requirements.txt
Fiona==1.9.3
numpy==1.23.5
numpy==1.24.3
pandas==1.5.3
pandas==2.0.1
progress==1.6
pymannkendall==1.4.3
PyYAML==6.0
PyYAML==6.0.1
rasterio==1.3.6
scikit_learn==1.2.2
scipy==1.10.1
Shapely==2.0.1
